rteval: Add __contains__ in rtevalConfig
Add the __contains__ function to the rtevalCfgSection class  to make "in"
function correctly.

For example in cyclictest.py, self.__cfg is not a dictionary, it is an
instance of the rtevalCfgSection class, therefore
key in self.__cfg
does not work as expected. In order to make sure this is fixed
everywhere, instead of comparing to self.__cfg.keys() everywhere this is
used, implement the __contains__ function so that "in" works as
expected.

This bug was introduced by the commit fd3b732f714d ("rteval: 2to3
transformations")

Reported-by: Atsushi Nemoto <atsushi.nemoto@sord.co.jp>
Signed-off-by: John Kacur <jkacur@redhat.com>
Correction
Suggested-by: Jeff Epler <jepler@unpythonic.net>
Signed-off-by: John Kacur <jkacur@redhat.com>