rteval: osinfo.py: Fix RT kernel detection
Since the osinfo.py module was developed, RT kernels have moved on
from identifying as "RT" to "PREEMPT_RT". This change in the kernel
identifier causes osinfo.py to incorrectly detect RT kernels as
non-RT.

Update the check for detecting RT kernel to be able to handle both the
variants.

Signed-off-by: Punit Agrawal <punit1.agrawal@toshiba.co.jp>