rteval: services.py: Fix incorrect detection of container environment
rteval mistakenly detects that it is running inside a container even
though it is running directly on the host. On further investigation
this was found to be due to change in behaviour around byte strings
and strings when going from python2 to python3.

In python3 byte strings are not equivalent to strings, i.e., b'' == ''
is False. The string comparison functions in services.py are still
relying on the old behaviour in python2 where they were equivalent.

Update the byte string processing by converting them to string.

Signed-off-by: Punit Agrawal <punit1.agrawal@toshiba.co.jp>